apex/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic[email]>=2.7.0
pydantic-settings>=2.2.1
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
redis>=5.0.4
anthropic>=0.28.0
qdrant-client>=1.9.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.7
httpx>=0.27.0
structlog>=24.1.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
tenacity>=8.3.0
aiofiles>=23.2.1
greenlet>=3.0.3
